+++ src/Enums/CurrencyEnum.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FurkanMeclis\PayTRLink\Enums;
+
+enum CurrencyEnum: string
+{
+    case TL = 'TL';
+    case USD = 'USD';
+    case EUR = 'EUR';
+    case GBP = 'GBP';
+}