+++ src/Enums/LinkTypeEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace FurkanMeclis\PayTRLink\Enums;
+
+enum LinkTypeEnum: string
+{
+    case Product = 'product';
+    case Collection = 'collection';
+}