+++ src/Exceptions/PayTRException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FurkanMeclis\PayTRLink\Exceptions;
+
+use Exception;
+
+class PayTRException extends Exception
+{
+    //
+}