+++ src/Facades/PayTRLink.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace FurkanMeclis\PayTRLink\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+ * @see \FurkanMeclis\PayTRLink\PayTRLinkService
+ */
+class PayTRLink extends Facade
+{
+    protected static function getFacadeAccessor(): string
+    {
+        return \FurkanMeclis\PayTRLink\PayTRLinkService::class;
+    }
+}