+++ tests/TestCase.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace VendorName\Skeleton\Tests;
+namespace FurkanMeclis\PayTRLink\Tests;
 
+use FurkanMeclis\PayTRLink\PayTRLinkServiceProvider;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Orchestra\Testbench\TestCase as Orchestra;
-use VendorName\Skeleton\SkeletonServiceProvider;
 
 class TestCase extends Orchestra
 {
@@ -13,20 +13,26 @@ class TestCase extends Orchestra
         parent::setUp();
 
         Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'VendorName\\Skeleton\\Database\\Factories\\'.class_basename($modelName).'Factory'
+            fn (string $modelName) => 'FurkanMeclis\\PayTRLink\\Database\\Factories\\'.class_basename($modelName).'Factory'
         );
     }
 
     protected function getPackageProviders($app)
     {
         return [
-            SkeletonServiceProvider::class,
+            PayTRLinkServiceProvider::class,
         ];
     }
 
     public function getEnvironmentSetUp($app)
     {
         config()->set('database.default', 'testing');
+
+        // Set PayTR config for testing
+        config()->set('paytr-link.merchant_id', 'test_merchant_id');
+        config()->set('paytr-link.merchant_key', 'test_merchant_key');
+        config()->set('paytr-link.merchant_salt', 'test_merchant_salt');
+        config()->set('paytr-link.debug_on', 1);
 
         /*
          foreach (\Illuminate\Support\Facades\File::allFiles(__DIR__ . '/../database/migrations') as $migration) {